requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

fastapi==0.108.0
starlette==0.31.1
requests==2.31.0
numpy==1.24.0
SQLAlchemy==1.4.36
pika==1.3.2
PyYAML==6.0.1
pyarrow==16.1.0
PyMySQL==1.1.1
pandas==2.2.2
torch==2.5.0
transformers==4.48.3
cryptography==42.0.8
pymilvus==2.4.4
psycopg2-binary==2.9.9
gunicorn==22.0.0
sentence-transformers==3.0.1
gguf==0.6.0
sentencepiece==0.2.0
uvicorn==0.23.2
unsloth[cu124-torch250] @ git+https://github.com/unslothai/unsloth.git@027b5100c7492bce025fe2f03681f440843e4666
bitsandbytes==0.44.1
unsloth-zoo<2025.2.6
xformers==0.0.28.post2
peft==0.14.0
celery[redis]==5.4.0
redis==5.2.1
aiofiles==24.1.0
pydantic==2.10.3
python-multipart==0.0.20
PyYAML==6.0.1
protobuf==3.20.0
tiktoken==0.7.0
sentencepiece==0.2.0
triton==3.1.0
PyMuPDF==1.25.3